fixup! WIP: Moving UserStorage, some commented out code in UserCredentialStoreManager and MapUserProvider
