+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo8_0_0.java
