+++ server-spi-private/src/main/java/org/keycloak/storage/ExportImportManager.java
@@ -1,5 +1,6 @@
 package org.keycloak.storage;
 
+import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
 import org.keycloak.representations.idm.RealmRepresentation;
 
@@ -7,5 +8,7 @@ import org.keycloak.representations.idm.RealmRepresentation;
  * @author Alexander Schwartz
  */
 public interface ExportImportManager {
-    void importRealm(RealmRepresentation rep, RealmModel newRealm);
+    void importRealm(RealmRepresentation rep, RealmModel newRealm, boolean skipUserDependent);
+
+    void updateRealm(RealmRepresentation rep, RealmModel realm);
 }
