+++ server-spi-private/src/main/java/org/keycloak/storage/MigrationManager.java
@@ -4,8 +4,12 @@ import org.keycloak.models.RealmModel;
 import org.keycloak.representations.idm.RealmRepresentation;
 
 /**
+ * Handle the migration of the datastore and an imported realm representation.
+ * Will eventually be handled by the store directly.
+ *
  * @author Alexander Schwartz
  */
+@Deprecated
 public interface MigrationManager {
 
     void migrate();
